Cook torrance finalized

+++ raytracer.h
@@ -28,11 +28,11 @@ public:
 
     glm::vec3 L( Ray, IntersectionRecord, size_t );
 
-    double rSchlick1(const glm::vec3 normal, glm::vec3 incident, double n1, double n2);
+    double rSchlick1( const glm::vec3 normal, glm::vec3 incident, double n1, double n2);
 
-    double rSchlick2(const glm::vec3 normal, glm::vec3 incident, double n1, double n2);
+    double rSchlick2( const glm::vec3 normal, glm::vec3 incident, double n1, double n2);
 
-    glm::dvec3 cook_torrance( glm::vec3 wi, glm::vec3 wo, IntersectionRecord intersection_record );
+    glm::dvec3 cook_torrance( glm::dvec3 wi, glm::dvec3 wo, IntersectionRecord intersection_record );
 
 private:
